Fixing clang 64bit to 32 conversion warning

+++ src/core/ext/filters/fault_injection/fault_injection_filter.cc
@@ -119,7 +119,7 @@ absl::StatusOr<FaultInjectionFilter> FaultInjectionFilter::Create(
 }
 
 FaultInjectionFilter::FaultInjectionFilter(ChannelFilter::Args filter_args)
-    : index_(grpc_channel_stack_filter_instance_number(
+    : index_((int)grpc_channel_stack_filter_instance_number(
           filter_args.channel_stack(),
           filter_args.uninitialized_channel_element())),
       service_config_parser_index_(
