+++ src/core/ext/xds/file_watcher_certificate_provider_factory.cc
@@ -131,7 +131,7 @@ FileWatcherCertificateProviderFactory::CreateCertificateProvider(
       file_watcher_config->private_key_file(),
       file_watcher_config->identity_cert_file(),
       file_watcher_config->root_cert_file(),
-      file_watcher_config->refresh_interval().millis() / GPR_MS_PER_SEC);
+      (unsigned int)(file_watcher_config->refresh_interval().millis() / GPR_MS_PER_SEC));
 }
 
 void FileWatcherCertificateProviderInit() {
